[BOOK-170] feat: vm 구현(추후 usecase 연결)
